コメント書き込み機能の関数作成
